hr - #7

+++ src/hr.API/Controllers/PeopleController.cs
@@ -1,13 +1,13 @@
-﻿using System;
-using AutoMapper;
-using Microsoft.AspNetCore.Mvc;
-using System.Collections.Generic;
-using hr.Domain.Interfaces.Services;
+﻿using AutoMapper;
 using hr.API.ViewModels;
-using System.Threading.Tasks;
+using hr.Domain.Interfaces.Services;
 using hr.Domain.Models.Entities;
 using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+using System;
+using System.Collections.Generic;
 using System.Net.Mime;
+using System.Threading.Tasks;
 
 namespace hr.API.Controllers
 {
@@ -26,17 +26,58 @@ namespace hr.API.Controllers
 
         // GET: api/people
         [HttpGet]
-        public async Task<IEnumerable<PeopleViewModel>> GetAsync()
-        {            
-            return _mapper.Map<IEnumerable<PeopleViewModel>>(await _peopleService.GetAll());
+        [Consumes(MediaTypeNames.Application.Json)]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+        public async Task<IActionResult> GetAsync()
+        {
+            try
+            {
+                var entities = _mapper.Map<IEnumerable<PeopleViewModel>>(await _peopleService.GetAll());
+
+                if (entities == null)
+                {
+                    return NotFound();
+                }
+
+                return Ok(entities);
+
+            }
+            catch (Exception ex)
+            {
+                return ApiError500(ex);
+            }
 
         }
 
         // GET api/people/5
         [HttpGet("{id:guid}")]
-        public async Task<PeopleViewModel> GetAsync(Guid id)
+        [Consumes(MediaTypeNames.Application.Json)]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        public async Task<IActionResult> GetAsync(Guid id)
         {
-            return _mapper.Map<PeopleViewModel>(await _peopleService.GetById(id));
+
+            try
+            {
+                var entity = _mapper.Map<PeopleViewModel>(await _peopleService.GetById(id));
+
+                if (entity == null)
+                {
+                    return NotFound();
+                }
+
+                return Ok(entity);
+
+            }
+            catch (Exception ex)
+            {
+                return ApiError500(ex);
+            }
         }
 
         // POST api/people
@@ -44,42 +85,108 @@ namespace hr.API.Controllers
         [Consumes(MediaTypeNames.Application.Json)]
         [ProducesResponseType(StatusCodes.Status201Created)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<IActionResult> PostAsync([FromBody] PeopleViewModel people)
         {
-            if(people == null)
+            try
             {
-                return BadRequest();
+
+                if (people == null || !ModelState.IsValid)
+                {
+                    return BadRequest();
+                }
+
+                var addedPeople = await _peopleService.Add(_mapper.Map<People>(people));
+                var addedPeopleVM = _mapper.Map<PeopleViewModel>(addedPeople);
+
+                return ApiCreated201(addedPeopleVM);
+
             }
-
-            var addedPeople = await _peopleService.Add(_mapper.Map<People>(people));
-
-            return new ObjectResult(addedPeople) { StatusCode = StatusCodes.Status201Created };
+            catch (Exception ex)
+            {
+                return ApiError500(ex);
+            }
         }
 
         // PUT api/people/5
         [HttpPut("{id:guid}")]
-        public async Task PutAsync(Guid id, [FromBody] PeopleViewModel people)
+        [Consumes(MediaTypeNames.Application.Json)]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        public async Task<IActionResult> PutAsync(Guid id, [FromBody] PeopleViewModel people)
         {
-            await _peopleService.Set(_mapper.Map<People>(people));
+            try
+            {
+                var entity = await _peopleService.GetById(id);
+
+                if (entity == null)
+                {
+                    return NotFound();
+                }
+
+                if (!ModelState.IsValid)
+                {
+                    return BadRequest();
+                }
+
+                await _peopleService.Set(_mapper.Map<People>(people));
+
+                return ApiOkNoContent204();
+
+            }
+            catch (Exception ex)
+            {
+                return ApiError500(ex);
+            }
         }
 
         // DELETE api/people/5
         [HttpDelete("{id:guid}")]
+        [Consumes(MediaTypeNames.Application.Json)]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
         public async Task<IActionResult> DeleteAsync(Guid id)
         {
             try
             {
                 var entity = await _peopleService.GetById(id);
 
+                if (entity == null)
+                {
+                    return NotFound();
+                }
+
                 await _peopleService.Remove(entity);
 
-                return new ObjectResult(null) { StatusCode = StatusCodes.Status204NoContent };
+                return ApiOkNoContent204();
 
             }
-            catch(Exception ex)
+            catch (Exception ex)
             {
-                return BadRequest();
-            }            
+                return ApiError500(ex);
+            }
+
+        }
+
+        private ObjectResult ApiError500(Exception ex)
+        {
+            //log the exception ex
+            return StatusCode(500, new { errorMessage = "Contact the support." });
+        }
+
+        private ObjectResult ApiCreated201(PeopleViewModel people)
+        {
+            return StatusCode(201, people);
+        }
+
+        private ObjectResult ApiOkNoContent204()
+        {
+            return new ObjectResult(null) { StatusCode = StatusCodes.Status204NoContent };
         }
     }
+
 }
